refactor(higher-order): split 1647-line file by responsibility; remove unused alias kind
Split rust/src/interpreter/higher-order-operations.rs into a higher_order/
submodule whose files are organized by responsibility, all under the §14.1
500-line limit:

  common.rs        ExecutableCode + extraction/execution helpers
  fast_kernels.rs  fast-path map/predicate/fold kernel detection
  runners.rs       quantized + plain kernel runners
  hedged.rs        hedged race executors (map/predicate/fold)
  map.rs           op_map
  filter.rs        op_filter
  any.rs / all.rs / count.rs  op_any / op_all / op_count

Behavior is preserved: 759 lib tests + 57 integration tests pass.

Also remove the dead CoreWordAliasKind::Deprecated variant. The
CORE_WORD_ALIASES table contains only canonical sugar/symbol/input-helper
entries (per SPECIFICATION.md §6/§7.0); deprecated migration aliases live
separately in interpreter/deprecated-core-aliases.rs. Removing the variant
makes that separation enforceable at the type level.
