+++ rust/src/interpreter/higher_order/mod.rs
@@ -1,0 +1,25 @@
+mod all;
+mod any;
+mod common;
+mod count;
+mod fast_kernels;
+mod filter;
+mod hedged;
+mod map;
+mod runners;
+
+pub(crate) use common::{
+    execute_executable_code, extract_executable_code, extract_predicate_boolean, ExecutableCode,
+};
+pub(crate) use hedged::{
+    execute_hedged_fold_kernel, execute_hedged_map_kernel, execute_hedged_predicate_kernel,
+};
+pub(crate) use runners::{
+    execute_quantized_fold_kernel, execute_quantized_map_kernel, execute_quantized_predicate_kernel,
+};
+
+pub use all::op_all;
+pub use any::op_any;
+pub use count::op_count;
+pub use filter::op_filter;
+pub use map::op_map;
